+++ Bin/ClearFolders.py
@@ -1,4 +1,4 @@
-#Authored by mostly nick :)
+# Authored by mostly Nick :)
 import os
 import shutil
 
@@ -8,6 +8,10 @@ def clear_folder(folder_path):
         for filename in os.listdir(folder_path):
             file_path = os.path.join(folder_path, filename)
             try:
+                # Skip deletion if the file is .gitignore
+                if filename == '.gitignore':
+                    continue
+                
                 if os.path.isfile(file_path) or os.path.islink(file_path):
                     os.unlink(file_path)  # Remove the file or link
                 elif os.path.isdir(file_path):
@@ -25,4 +29,4 @@ output_folder = 'Output'
 clear_folder(input_folder)
 clear_folder(output_folder)
 
-#print("Input and Output folders have been cleared!")
+# print("Input and Output folders have been cleared!")
